Removed old Types
